Move UfsUtils to client module, as it load an underfs and expose it as a tachyonFS client
